requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-SQLAlchemy
psycopg2
flask-marshmallow
Flask-WTF

# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn
